requirements.txt
spacy
gtts
openpyxl
pyspellchecker